requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==7.1.2
coverage==7.3.2
dnspython==2.4.2
exceptiongroup==1.1.3
Flask==1.1.4
flask-restplus==0.13.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
MarkupSafe==1.1.1
packaging==23.2
pkgutil_resolve_name==1.3.10
pluggy==1.3.0
pymongo==4.6.0
pytest==7.4.3
pytest-mock==3.12.0
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.0
six==1.16.0
tomli==2.0.1
urllib3==2.1.0
Werkzeug==0.16.1
zipp==3.17.0
